Fixed issues with text preprocessing
In IWSPA there were emails with multipart/mixed, alternative, related parts which included the text for some reason (almost 80-90% of the emails). Therefore, to fix that, we include these types if they are found in the *payload dictionary* data structure that we use which is a dictionary as follows:
mime-type : contents

This data structure does not include multipart parts if they are actually multipart (i.e. they include other mimetypes/parts under them), but it does include multipart parts if they are a leaf of the emails tree structure.